+++ src/components/FhiHeader/__tests__/FhiHeader.spec.tsx
@@ -11,6 +11,7 @@ describe('FhiHeader', () => {
           className="test"
           projectName="Testproject"
           logo="../assets/logo-norsk-kun-forkortelse-svart-uten-luft-rundt.svg"
+          logoLight="../assets/logo-norsk-kun-forkortelse-hvit-uten-luft-rundt.svg"
           menuItems={[
             {
               name: 'Dette er en tittel',