+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 
 
+## Unreleased
+
+> May 04, 2022
+
+* :tada: **Enhancement** Added prop for iconLight in `FhiHeader`
+* :tada: **Enhancement** Fixed margin/padding bottom on `FhiHeader`
+* :nut_and_bolt: **New** Added `FhiTabBar`
+
+
 ## 0.3.0
 
 > May 04, 2022
@@ -13,6 +22,7 @@
 
 * :tada: **Enhancement** Added disabled-option to `FhiButton`
 
+
 ## 0.1.0
 
 > Apr 29, 2022
@@ -23,6 +33,7 @@
 * :nut_and_bolt: **New** Added `FhiDetails` to package
 * :nut_and_bolt: **New** Added `FhiTheme` to package
 * :nut_and_bolt: **New** Added `FhiHeader` to package
+
 
 ## Examples